multi-user: avoid race with 3 or more users
The point in the loop where I replaced a check for dup for "this user" with
"okay for all users" is inside the loop while still reconciling users, so with
3 or more users it's possible for the default user to be the third or later
encountered, so the audit is too soon.

Since the `auditDupDefaults()` method is for _all_ users, we don't need to call
it inside the loop and can just call it once at the end.  By which point we've
seen all the users and know that we're seeing a complete set.